Fix if in line 1049 and comment in line 1047

+++ core.sh
@@ -9,7 +9,7 @@ _CAT()
 {
     # Tag para sinalizar que precisa parar.
     local end_of_file='EOF'
-    
+
     INPUT=( "${@:-"%"}" )
     for i in "${INPUT[@]}"; do
         if [[ "$i" != "%" ]]; then
@@ -34,7 +34,7 @@ _GREP()
     # ou variável.
     local expression="$1"
     local receive="$2"
-    
+
     # Testando e buscando expressão.
     if [[ -z "$expression" ]]; then
         { echo 'MODULE _GREP ERROR. Not found variable $expression'; exit 1 ;}
@@ -54,7 +54,7 @@ _WC()
 {
     local check="$@" # Recebendo args
     local inc='0'    # Var incremento
-    
+
     for x in $check; do
         let inc++
     done
@@ -105,8 +105,8 @@ _NAME_FORMAT_PKG()
 _MANAGE_SCRIPTS_AND_ARCHIVES()
 {
     local packname="${1/%.slp/}"
-    local dir_desc="${dirlist/list/desc}"  
-    
+    local dir_desc="${dirlist/list/desc}"
+
     if ! [[ -e "/info/desc" ]]; then
         echo -e "${red}[ERROR!]${end} /info/desc does not exist. ABORT!"
         exit 1
@@ -159,13 +159,13 @@ _CREATE_LIST()
 {
     # Variáveis locais
     local packname="$1"
-    
+
     if ! tar tf "$packname" > "${dirlist}/${name_version_build}.list"; then
         echo -e "${red}[ERROR!]${end}\tNot Create ${dirlist}/${name_version_build}.list"
         return 1
     fi
     echo -e "${blue}[CREATE LIST]${end}\t On ${dirlist}/${name_version_build}.list"
-    return 0    
+    return 0
 }
 
 
@@ -177,7 +177,7 @@ _GENERATE_DESC()
     local DESC_PACKNAME="$1"
     local DESC_VERSION="$2"
     local DESC_BUILD="$3"
-    
+
     [[ ! -d "info" ]] && mkdir info # diretorio info não existe? crie.
     _CAT > "info/desc" << EOF
 ######################################################################
@@ -240,7 +240,7 @@ _GPG_SIGN()
 {
     local package="$1"
     local sig='sig'
-    
+
     # Pacote existe?
     if [[ ! -e "${package}.${format_pkg}" ]]; then
         echo "${red}[ERRO]${end} Unable to sign package. ${package}.${format_pkg}"
@@ -248,7 +248,7 @@ _GPG_SIGN()
         echo "For security reasons, do not pass the package on to third parties."
         return 1
     fi
-    
+
     # Gerando Assinatura no pacote
     gpg --detach-sign --pinentry-mode loopback "${package}.${format_pkg}" &>/dev/null || \
     gpg --detach-sign "${package}.${format_pkg}" || return 1
@@ -266,7 +266,7 @@ _VERIFY_ON()
   ( # Subshell marota
     local dir_info='info'  # Diretorio info que contem informações como (desc)
     local info_desc='desc' # Descrição do pacote
-    
+
     if [[ ! -d "$dir_info" ]]; then # Diretório info existe?
         echo -e "${red}[ERROR!]${end} ${pink}${dir_info}${end} directory\n"
         echo -e "It's necessary your package have the DIRECTORY ${pink}info${end}."
@@ -338,7 +338,7 @@ _VERIFY_ON()
    done
 
  )
- 
+
 _SUBSHELL_STATUS
 }
 
@@ -346,9 +346,9 @@ _LIST_ARCHIVES_DIRECTORIES()
 {
     local packname="${1}.list"
     local LIST_CLEAN_DIRECTORIES
-    
+
     echo -e "${cyan}[Clean]${end} List: ${dirlist}/${packname}"
-    
+
     # Apagando coisas desnecessárias na lista e substituindo.
     sed -i "
         s/^\.\///g
@@ -386,7 +386,7 @@ _LIST_ARCHIVES_DIRECTORIES()
         /^var\/lib\/spm\/list/d
         /^var\/lib\/spm\/remove/d
     " "${dirlist}/${packname}"
-    
+
     # Lista de diretórios para usar no loopzinho
     LIST_CLEAN_DIRECTORIES=(
         'var'
@@ -412,7 +412,7 @@ _LIST_ARCHIVES_DIRECTORIES()
         elif [[ "$view" =~ ^${LIST_CLEAN_DIRECTORIES[3]}/(bin|etc|lib|lib\/(pkgconfig)|local|share|doc|include|libexec|sbin|src)$ ]]; then
             local view="${view//\//\\/}"
             sed -i "/^$view$/d" "${dirlist}/${packname}"
-        elif [[ "$view" =~ ^${LIST_CLEAN_DIRECTORIES[4]}/(keymaps|fonts|pixmaps|applications|doc|man|man\/man[[:digit:]]+|man\/.{2})$ ]]; then      
+        elif [[ "$view" =~ ^${LIST_CLEAN_DIRECTORIES[4]}/(keymaps|fonts|pixmaps|applications|doc|man|man\/man[[:digit:]]+|man\/.{2})$ ]]; then
             local view="${view//\//\\/}"
             sed -i "/^$view$/d" "${dirlist}/${packname}"
 		elif [[ "$view" =~ ^${LIST_CLEAN_DIRECTORIES[5]}/(www|httpd)$ ]]; then
@@ -476,7 +476,7 @@ _INSTALL_PKG()
     local name_version_build tmp_pack
     local tmp_pack
     local PRE_SH='pre.sh'
-    
+
     # Descompactando desc primeiro para exibir informações do pacote.
     # e carregando o arquivo desc do programa ;)
     if ! tar xf "${packname}" -C "/tmp/" "./${descme}"; then
@@ -491,10 +491,10 @@ _INSTALL_PKG()
         { kill $pid; wait $pid 2>/dev/null; echo ;}
         return 1
     fi
-    
+
     { kill $pid; wait $pid 2>/dev/null; echo ;}
-    
-    
+
+
     # Variavel puxando de source...
     # Está variavel é importante se caso
     # o usuario passe um caminho completo do pacote exemplo:
@@ -520,7 +520,7 @@ _INSTALL_PKG()
         else
             echo -e "${red}[Pre-Installation]${end} Cannot extract ${PRE_SH}, ABORT"
             return 1
-        fi  
+        fi
         bash "/tmp/info/$PRE_SH"
         tar xf "${packname}" -C / || return 1
         echo -e "${blue}[EXTRACT]${end}\t On Your root, OK."
@@ -533,11 +533,11 @@ _INSTALL_PKG()
         _CREATE_LIST "$1" || return 1 # Criando lista
         _MANAGE_SCRIPTS_AND_ARCHIVES "${name_version_build}" || return 1
     fi
-    
+
     # Ajustando lista.
     _LIST_ARCHIVES_DIRECTORIES "${name_version_build}"
  ) # Fim da subshell suicide
- 
+
 _SUBSHELL_STATUS
 }
 
@@ -690,7 +690,7 @@ _PRE_REMOVE()
                  search_pack="${search_pack/%.desc/}" # Cortando para impressão.
                  echo -e "${red}[FOUND]${end} $search_pack"
                  continue
-             
+
              fi
         else
             inc="$(( $inc + 1 ))"
@@ -727,7 +727,7 @@ _REMOVE_NOW()
         if [[ -e "${dirremove}/${packname}.rm" ]]; then
             rm "${dirremove}/${packname}.rm" && echo -e "${blue}[REMOVED]${end}\t${packname}.rm\n"
         fi
-    fi    
+    fi
 
     ###########################
     # Removendo:
@@ -735,7 +735,7 @@ _REMOVE_NOW()
     # Links simbólicos
     # Diretórios vazios
     ###########################
-    
+
     while IFS= read thefile; do
         if [[ -f "$thefile" ]]; then
             rm "$thefile" &>/dev/null && echo -e "Delete\t${thefile}"
@@ -758,11 +758,11 @@ _REMOVE_NOW()
     # Removendo diretórios vazios, e forçando para
     # toda hierarquia do pacote seja removido.
         while IFS= read thedir; do
-            if [[ -d "$thedir" ]] && [[ -z "$(ls -A ${thedir})" ]]; then 
+            if [[ -d "$thedir" ]] && [[ -z "$(ls -A ${thedir})" ]]; then
                   rmdir -p "${thedir}" &>/dev/null && echo -e "Delete\t${thedir}"
             fi
     done < "${dirlist}/${packname}.list"
-    
+
     echo -e "\nClean archives .desc and .list\n"
 
     # Removendo lista e Descrição
@@ -772,7 +772,7 @@ _REMOVE_NOW()
             list)   DIREC="$dirlist";;
         esac
         if [[ -e "${DIREC}/${packname}.${removeitem}" ]]; then
-            if rm "${DIREC}/${packname}.${removeitem}"; then 
+            if rm "${DIREC}/${packname}.${removeitem}"; then
                 echo -e "${blue}[REMOVE]${end} ${DIREC}/${packname}.${removeitem} SUCCESSFULLY"
             else
                 echo -e "\n${red}[ERROR]${end} It was not possible remove ${DIREC}/${packname}.list"
@@ -821,7 +821,7 @@ _SEARCH_PKG()
         elif [[ "$searchpkg" =~ ${re} ]] && [[ "$ONE_LINE" = '1' ]]; then
             inc=$(( $inc + 1 )) # Incrementando
             source $searchpkg # Carregando informações do pacote para impressão.
-            echo -e "${cyan}[FOUND]${end} ${pkgname}-${version}-${build}"       
+            echo -e "${cyan}[FOUND]${end} ${pkgname}-${version}-${build}"
         fi
     done
     if [[ "$inc" -eq '0' ]]; then
@@ -878,27 +878,24 @@ _UPDATE_SPM()
     local connectivity="${red}[FAILED]${end}"
     local tmp_dir_spm="/tmp/${PRG}pkg"
     local DOWNLOAD m
-    
+
     # Verifica conexão com internet usando networkmanager
     if type -P nmcli &>/dev/null; then
         [[ "$(nmcli -g connectivity g)" = "full"  ]] && connectivity="${cyan}[OK]${end}"
         echo -e "Internet\t$connectivity"
     fi
 
-    # Verificar se /tmp/spm já existe
-
-    files=$(shopt -s nullglob dotglob; echo /tmp/spm/*)
-    if (( ${#files} ))
-    then
-	rm -rf /tmp/spm/;
+    # Verifica se /tmp/spm já existe
+    if [ -d "/tmp/spm/" ]; then
+	     rm -rf /tmp/spm/;
     fi
 
-    
+
     # Ok, Puxe o repositorio agora!
     pushd /tmp &>/dev/null
     git clone "$link" || return 1
     pushd "${tmp_dir_spm}" &>/dev/null
-    
+
     # Dando permissões e copiando arquivos para seus lugares.
     echo -e "\nPermission and Copy archives\n"
     for m in "${PRG}.conf" "${PRG}.8" 'core.sh' 'help.sh' "$PRG"; do
@@ -922,6 +919,6 @@ _UPDATE_SPM()
              (core.sh|help.sh) cp -v "$m" "/usr/libexec/spm/" || return 1 ;;
         esac
     done
-    
+
     [[ -d "${tmp_dir_spm}" ]] && rm -r "${tmp_dir_spm}"
 }
